Move CFGToSCFPass
