+++ mlir/CMakeLists.txt
@@ -40,12 +40,13 @@ add_subdirectory(include/mlir-extensions/Dialect/gpu_runtime/IR)
 
 set(SOURCES_LIST
     lib/Conversion/SCFToAffine/SCFToAffine.cpp
+    lib/Conversion/cfg_to_scf.cpp
     lib/Conversion/gpu_runtime_to_llvm.cpp
     lib/Conversion/gpu_to_gpu_runtime.cpp
     lib/Conversion/util_to_llvm.cpp
     lib/Dialect/gpu_runtime/IR/gpu_runtime_ops.cpp
-    lib/Dialect/plier/dialect.cpp
     lib/Dialect/imex_util/dialect.cpp
+    lib/Dialect/plier/dialect.cpp
     lib/Transforms/arg_lowering.cpp
     lib/Transforms/call_lowering.cpp
     lib/Transforms/canonicalize_reductions.cpp
@@ -74,12 +75,13 @@ set(SOURCES_LIST
     )
 set(HEADERS_LIST
     include/mlir-extensions/Conversion/SCFToAffine/SCFToAffine.h
+    include/mlir-extensions/Conversion/cfg_to_scf.hpp
     include/mlir-extensions/Conversion/gpu_runtime_to_llvm.hpp
     include/mlir-extensions/Conversion/gpu_to_gpu_runtime.hpp
     include/mlir-extensions/Conversion/util_to_llvm.hpp
     include/mlir-extensions/Dialect/gpu_runtime/IR/gpu_runtime_ops.hpp
-    include/mlir-extensions/Dialect/plier/dialect.hpp
     include/mlir-extensions/Dialect/imex_util/dialect.hpp
+    include/mlir-extensions/Dialect/plier/dialect.hpp
     include/mlir-extensions/Transforms/arg_lowering.hpp
     include/mlir-extensions/Transforms/call_lowering.hpp
     include/mlir-extensions/Transforms/canonicalize_reductions.hpp
